[KGEN] Land the new elaborator. (#7432)
The new elaborator currently works on all the places we call kgen-opt, but doesn't work on all the kernels. There are some bugs to track down, so it's currently off for anything that uses the `kgen` driver, but on for anything using the `kgen-opt` driver.

The new elaborator explicitly builds the expansion tree while it runs elaboration, and makes an effort to limit global information sharing. This design will allow us to take advantage of parallelizing and caching sub-trees of elaboration.

MODULAR_ORIG_COMMIT_REV_ID: dacade3f99656e17b88f57c49b7e3cf298c79483