[Kernels] Change matmul to work on mixed precision dtypes (#7601)
Before the input dtypes had to match the output dtypes, but with this
change the output dtype does not have to match the input dtypes. Within
the innermost micro-kernel we cast the input dtype to the output dtype
and perform the fma using the output dtypes' precision. This would
allow us to do matmul where the output and computation is performed
in i32 but the input data is in i8 for example.

LINALG_ORIG_REV_ID: f11fcbb48a899efad7b68cc63de9cdfae66960d8