[Lit] Allow inplace operations to take by-val or by-ref self. (#7689)
Some struct types are conceptually reference semantic, and thus
in-place operations don't have to modify the struct itself, they
can modify what the struct points to.  e.g. we may want this to
work some day:

```
  let x = somePythonFunction()
  x += 1  # mutates the pointee not the pointer.
```

This patch relaxes the by-ref checking to allow either by-ref or
by-val.  We may eventually want to eliminate this entirely as being
not-worth-checking for, but this is a logical step forward.

This fixes #7573

MODULAR_ORIG_COMMIT_REV_ID: 662442ea5b60f68840e844d9bcf01b76c48a4077