[KGEN] Remove ParamDeclAttr from ParamBindAttr (NFC) (#7730)
And just use a StringAttr instead. In the future, the presence of
ParamDeclAttr will itself indicate a param declaration, so it should not
be included in parameter binding lists. Also, the type of the decl is
stored redundantly.

MODULAR_ORIG_COMMIT_REV_ID: 64b6202eaa1c46d8d424187291c400e993da3278