[Lit] Introduce explicit #lit.struct.extract/#pop.struct.extract attrs. (#7784)
This introduces explicit representation for struct.extract in the
meta-program, instead of using the emitMLIROperationCall mechanism.
The rationale here is that this is simpler and consistent with the
rest of the attributes used by the metaprogram, composes better with
fold hooks, and eliminates the need for the 'lift mlir' pass which has
some bugs.

MODULAR_ORIG_COMMIT_REV_ID: cbea60bdd922b1ee1a9292a98964892b329ce236