[Lit] Change `verifyBindings` to require a complete parameter spec. (#7807)
This changes verifyBindings to always require a full parameter spec
when computing a ParamBindArrayAttr.  Previously it was silently a
noop when passed no bindings, but this pushes that behavior into the
one place that can take this.

This is important because variadics means that we need to have a
specific place where we "complete" a parameter list: zero parameters
may either mean a reference to a symbol without bindings (this is valid
in a parameter expr context) or it may mean that the variadic list is
bound to zero values (this happens in a value context).

MODULAR_ORIG_COMMIT_REV_ID: e0c7cd46ea3dff5ca388cb82e24103a4e304c140