[Lit] Teach variadic parameter list to bind to empty list. (#7810)
This rearranges type checking of a parameter list to proceed in order
of the parameter declarations, using up binding values on demand. This
teaches the last variadic parameter that it can bind to an empty list,
and teaches DRValue conversion of a signature typed value to perform
binding of the (missing) parameters.  This ensures that varargs and
(someday) default arguments get bound at this point.

MODULAR_ORIG_COMMIT_REV_ID: 5808c5c2abae017091d327ba7d7f224e4cd37168