[KGEN] Fix non-determinism in ParameterUseDefGraph usesFromAbove (#7805)
In signature types, the declarations of parameters must satisfy the
property that a use of a parameter in the declaration list can only
occur after that parameter is declared in the list. A use of a parameter
in a parameter declaration can only occur in its type. That means that when
outlining occurs, the outlined closure must order the captured
parameters topologically. The parameter use def graph does a post-order
traversal and checks the type for uses first, that means it does the
sort automatically, but because we were using a SmallPtrSet, it can get
reordered nondeterministically and result in crashes.

This fixes that by making it a SetVector...

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: f46c6ee0b269126e0cb6674c783d0b28b5b4d720