[Cache] Make the region hash parametrizable on any attribute. (#7782)
We know we want the region hashes parametrized on constant hashes and symbols, but there may be other attributes in the future. This change allows us to take any attribute, have it implement the appropriate interfaces, and use it to parametrize cached IR on that attribute.

MODULAR_ORIG_COMMIT_REV_ID: f53670d3417341846d349456e99487e97659da3e