[Kernels] Add a DynamicRankBuffer to simplify how we interact with C++ (#7870)
One of the long standing issues is that we do not have a
rank-polymorphic type withing lit (or the mlir code for that matter).
This was done because a rank-polymorphic type would encure heavy cost
during indexing. As a result, to emulate dynamic rank we had to pass in
the NDBuffer components by arguments from C++ and do a rank dispatch
within the function. Because of this, the functions took a gazillion
arguments, not consistent (some were doing the rank dispatch in C++
land), were not easily generalizable, and create very long function
names.

This introduces a DynamicRankBuffer type which allows you to grab a
Buffer and NDBuffer from it. Code is shared and is simplifies and
localizes to a single place.

BUFFER_ORIG_REV_ID: a58765a88db5f288a03000f96ceaa41e7a7275fd