[lit] Fixes a bug in the out-of-place transpose (#7770)
Given an input 2x3 input matrix the buggy code would read pairs (0,2)
and (2,0) and write pairs (2,0) and (0,2). But (2,0) is not an element
of the input matrix and (0,2) is not an element of the output matrix.

For the out-of-place transpose there is no point in swapping pairs for
optimization.

LINALG_ORIG_REV_ID: 7c98e1d62d2741b0178f81f15b3d027947d1b9c4