[Lit] Wire subscript syntax up to the `__getitem__` dunder method. (#7917)
This allows type authors to implement the __getitem__ method to
enable values to be subscripted.  This is an extended version of
the Python semantics (given we support overloading) that allows
you to define N indices instead of a single version that takes a
tuple (also convenient because we don't have tuples yet).

Note that this has a very very important limitation: subscripts are
NOT wired up to __setitem__ yet.  This means that you can read values
with `.. = v[i]` but you cannot store to them with `v[i] = ..`.  For
this, please continue to call __setitem__ directly.

MODULAR_ORIG_COMMIT_REV_ID: 32bf97f137803a2ec3cdcb258f00d5ab14958b48