[Lit] Diagnose unused expression results with a warning. (#8019)
It is a common error to forget to use values, to accidentally forget
to use a function result, or to forget the parens on a function call.
We have a number of these in-tree already (filed at Issue #8018).

We now diagnose this common error with a warning that has a fixit
hint with a recommended remedy, e.g.:

```
x.lit:13:4: warning: '__mlir_type.index' value is unused
  4+4
  ~^~
  _ =
x.lit:15:3: warning: function pointer was formed but not called, did you forget '()'s?
  test
  ^~~~
      ()
```

Someday you'll be able to mash a VSCode button to auto-apply those
for you.  In the meantime this will help eliminate a class of bugs.

MODULAR_ORIG_COMMIT_REV_ID: 68629b8a1859c83de8c032b62ae9a1069fcce1f5