[LIT] Add a basic lit language server (#8009)
This is mostly just a skeleton and setups up enough to give
diagnostics to the user (fixits will be added in a followup),
but not much more than that. This should give us a nice base to
start layering on language functionality on top of. We need to
fix some aspects of lit before then, though, such as the
compilation model (to avoid super super slow IDE feedback).

MODULAR_ORIG_COMMIT_REV_ID: 4d3b093b743239e332d9446576f6af07afb1a986