[KGEN] Force CompilerRT symbols to link into executable/so. (#8045)
When a library links CompilerRT it will see that none of the symbols in
any of the .o files are actually used, and so the linker will just not
pull in anything. This patch adds an __attribute__((used)) function in
the CompilerRT header so that anything that includes that header will
pull in the object files. This is similar to `-Wl,-whole_archive` but
easier to make platform-agnostic.

Is what I'd like to say. Unfortunately, we have to manually register the
functions into the ORC cause it's being silly on linux platforms and so
the symbols in the current process are not being linked properly.

MODULAR_ORIG_COMMIT_REV_ID: ce2a510869bda3c3a1cfb15f955f9468b1252eb4