[lit-vscode] Add support for diagnostic fixits to the language server (#8087)
This lets us plug into the fixit infra we have in lit, and lets an IDE show
or automatically apply the correct fix for a diagnostic. The title of the
fixit is basically the same behavior as clangd.

This also fixes unused result warning to provide the correct range to the
fixit, as right now it will end up replacing text instead of inserting. This
was found during testing (which is much easier to see in an IDE).

MODULAR_ORIG_COMMIT_REV_ID: 3c895334ce85ca588aceca3f3c62951d992139ec