[lit-lang] Restructure module import to be more lazy (#7169)
We currently process module imports as soon as we see them,
even if we never actually reference the module itself. This PR
restructures import to instead create place holder decls for the
imports, and only resolve them if necessary (i.e. during name lookup,
where we know they actually get referenced).

Wildcard decls are a bit weird in this sense, given that we don't know
the full scope of what is supposed to be imported. To handle these,
when we encounter a name that doesn't correspond to a known decl
we resolve any lingering wild cards in that scope to see if they
populate the decl we are looking for.

MODULAR_ORIG_COMMIT_REV_ID: 4a3d4a1df3ca238b083eb8a92d93196148c5167d