[Profiling] Add initial version of trace types (#8152)
Previously each trace statement was associated with a level such that the trace was enabled if
level <= MODULAR_LLCL_MAX_PROFILING_LEVEL. However, this produced cluttered and large trace files
that weren't targeted for the types of traces a user was interested in. For e.g., it would be
reasonable to want to see detailed info on LIT kernels without seeing all of the details of the LLCL
runtime scheduling.

This PR adds the initial scaffolding for trace types so that traces can be logically grouped into
different bins, and the user can request to only see traces from a particular bin.

MODULAR_ORIG_COMMIT_REV_ID: 137d598b364c689e891dc3add54af7f1d61d8c0a