[Lit] Adjust call emission in preparation for removing `@nodebug_inline`. (#8216)
This changes the logic for emitting calls in parameter expressions and
when their operands are fully parameters to use the existing logic for
nodebug_inline when they are marked alwaysinline("nodebug") and switches
a few examples over.  This makes it so dynamic calls are emitted as
kgen.call, but we try to turn other things into folded parameters.

The rationale for this is that we're allowed to remove debug info from
these
things and we really do want things inlined in parameter expressions and
when
trivial to reduce IR size and ensure parameter canonicalization works.

MODULAR_ORIG_COMMIT_REV_ID: e6c9b7190e58bb14a483b32b37522e619661ed06