[Lit] Convert a bunch of stuff off nodebug_inline. (#8234)
This is generally straight-forward and mechanical at this point.
One thing that needs to be implemented is parameter folding support
for inlined functions in parameter expressions.

This is a reapply of PR8218 now that an elaborator bug is fixed.

MODULAR_ORIG_COMMIT_REV_ID: 0e71b539c93bd2eaef5d29b33ae6d1d7275c2568