[KGEN][Kernels] Fix async+multithread crash and reenable threading (#8224)
This PR fixes the race condition identified in
[Internal link]

Unfortunately, LLVM's coroutine lowering makes it impossible to avoid
the extra store when a coroutine reaches its final suspend. This means
the code to invoke the completion callback must be outside the
coroutine's resume function. All this callback code has been deleted
from the compiler and is performed by adding a check to
`_coro_resume_fn` in the standard library `Coroutine` type. This is
going to add a slight overhead, because now each resume of a coroutine
will have to check whether it's done and then, if it is, invoke the
completion callback, whereas before the compiler would simply insert the
call at the final suspends of the coroutine.

We will get that performance back when we implement coroutine handling
ourselves, and then some more.

This PR re-enables multithreading by default for `parallelForEachN` and
the threaded kernels.

Fixes #7968
Fixes [Internal link]

ALGORITHM_ORIG_REV_ID: cbae77d062d31b3f97648328f657653e9f4b3a9d