[lit-lang][stdlib] Actually for real fix async issues this time but no really (#8364)
The approach I took to fix the race condition by side-stepping the rogue
store generated by LLVM's coroutine splitting did not work because it
introduced another race condition when a coroutine schedules itself (via
`Task.__await__` and the schedule occurs before the coroutine exits and
checks its own completion. This reverts the coroutine lowering in LIT to
generate the call to the completion function, which means the only
plausible fix is a spin loop in the coroutine destructor...

This is a terrible solution mostly because it will waste cycles on the
main thread, but it will have to do.

MODULAR_ORIG_COMMIT_REV_ID: cd57ca76dc0105951a7c4d458cd58b3198919961