[KGEN] Teach the ParameterUseDefGraph and Elaborator about `kgen.param.if`. (#8435)
This patch adds support for `kgen.param.if` in the elaborator and the parameter graph, as well as tests for the same. This patch also makes sure to explicitly reject cases the elaborator would get confused on so we get a nice clean error.

MODULAR_ORIG_COMMIT_REV_ID: 36e1f5efe01b4239645ad7880951292ea75293c3