[LLCL] Don't call waiters inline within emplace (#8451)
Currently, we invoke all AsyncValue waiters when an AsyncValue is
emplaced, using the emplacer's call stack (and thread). There's two
issues:
1. A waiter may trigger another emplace, thus the stack depth is both
unbounded and not lexically obvious.
 2. RAII objects (eg for profiling) remain alive while waiters execute.

Here we implement a 'trampoline' to execute waiters on the emplacer's
thread, but only on the next go around the worker loop.

If the emplacer's thread is 'foreign' (ie not created by LLCL) then we
have a choice:
- (a) Add the waiter as an LLCL task. (However if the task queue is full
the task will be executed immediately, ie as per current behavior).
 - (b) Call the waiter directly, as per current behavior.

I've gone with (a) to address issue 2 in the common case. However
there's a risk folks might incorrectly assume waiters will only be run
after the current task has finished, which is unfortunately still not
true when the task queue is under heavy load. So we can switch using
kRunImmediatelyOnForeignThreads.

It seems only a few tests had the form:
```
   localVar = ...
   ref = AsyncValueRef<...>::allocate(...)
   ref.andThenSync([&localVar]() { ... })
   std::move(ref).emplace(...)
```
which now results in a dangling ref in the closure. I didn't audit the
whole codebase for that. In ASAN I trust.

The extra overhead is a virt call, emplacing into a TLS SmallVector,
looping over said vector, and some llvm::unique_function moves.

Reworked AsyncValueTest.cpp for more orthogonal coverage.

Fixes #7258.

MODULAR_ORIG_COMMIT_REV_ID: 2be40ceffcbace16e525caddec4b1446b5bbfff8