[KGEN] Fix `param.if` early return handling. (#8464)
The param.if op can use hlcf.return to perform an early return. This means that we need to handle it - and this actually allows us to avoid elaborating chunks of the function if the early return makes a lot of the code dead. This patch adds handling and testing for the early return case.

MODULAR_ORIG_COMMIT_REV_ID: 63c0150e0a983a42fea381d6ac3e266817b4787e