[lit-lang] Enhance `emitConditionValueAsI1` to preserve MValue'ness (#8476)
This enhances this function (used in if/while statements among others)
to emit the conversion to i1 as a parameter expression when the input
is a parameter expression.  This is important for "static if"s, because
they happen in dynamic compatible context, but want to preserve the
MValue'ness of the operand.  This should be no downside to doing this
always.

MODULAR_ORIG_COMMIT_REV_ID: 54920988ee65ec979c846e50ed686cc64b3fe9f5