[lit-lang] Further simplify/narrow inlineFunctionCallIntoMValue. (#8513)
This tightens this functionality up to only apply to functions whose
inputs are known constants and known *simple* constants.  This avoids
wasting compile time trying to fold things that won't actually be
constants, and aligns better with our goal of "inlining" constants
for the purposes of simplifying them - we don't want to actually build
a massive tree of parameter attrs.

The one casualty of this is that noop alwaysinline("nodebug") calls
like Bool.__bool__ are not being simplified away when the operand is
a DRValue. I don't think this is a bad thing though, since that has
nothing to do with the zero cost abstractions on parameters.

MODULAR_ORIG_COMMIT_REV_ID: 628528dc846a7231f0894f8fbff8f83e742f7b53