[lit-lang] Add support for `@parameter if`. (#8573)
This adds support for an explicitly marked if statement that requires
its condition to be a parameter expression, and uses that to guarantee
that only the true side of the if makes it past the elaborator.  This
provides a more bulletproof model than the previous automagic approach
that didn't really work.  This also reverts the other change to
`__lit_bool` emission.

MODULAR_ORIG_COMMIT_REV_ID: 3c669233d12683b477ce13dfdbb8fcb400a987d8