[Cache] Remove `CacheFindResult`. (#8565)
Using the UnknownLocationDecoder, we can slightly modify the BlobCache APIs to take an optional EncodedLocation. Between this and the UnknownLocationDecoder, we can now use the native error reporting in AsyncValue rather than our own special thing. This also simplifies the model quite a bit, *and* has the added bonus of being able to forward location information to cache errors.

The only slightly awkward thing is that `BlobCacheBackend::findImpl` now has to return the full `ErrorOr<std::optional<BufferRef>>`, but this is not a large burden for the simplification offered.

MODULAR_ORIG_COMMIT_REV_ID: 184e649450f780d9956a5b72dba111b24cf45863