[LLCL] Plumb time profile entries into addTask (#8630)
The simple idea here is to make the LLCL work queue the definitive
tracker of when a task enters the queue ('waiting' time), then
executes('execution' time). The profiling entry is passed directly so
that this is zero cost when disabled (at compile time). The profiling
entries are defined by a using and support the creation/restart/record
lifecycle directly.

However, the whole edifice started to come down around me so figured
better lock this in before going any further:
 - Got rid of most of the macros, only one left is for profiling scopes.
- Gave up supporting the 'suppress nested profiling entries' feature
when using TimeProfileScope. Honestly it should be done as a filter when
writing the trace, does not make sense to do it on the fly. However it
is still there if you use the explicit begin and end API.
- Make sure there's no surprised with detail-as-a-string-literal vs
detail-as-a-function-literal.

I think we should stop making the distinction between 'my trace type and
level have been enabled for the build' and 'profiling is active for this
run'. That would get rid of a whole layer of wrapper functions.

MODULAR_ORIG_COMMIT_REV_ID: 1fc887f3c74032b95933352e9be67fe83c3a23c7