[Telemetry] Remove use of telemetry for now (#8350)
Telemetry isn't getting actively used. It was a PoC exploration of
things in the C milestone, and is causing us some CMake slowdowns for
little value add since it's not getting used right now.

Specifically, we spend about 1 min 41 seconds each time we need to build
all of opentelemetry (and its dependencies) from scratch. Since the
CMake configure time is a serial process, this is unfortunately long and
is single-handedly the thing taking the longest.

Remove all use of Telemetry for now. We can bring back the library and
have it connect up to the Runtime for our initial needs when we're ready
to tackle that again.

MODULAR_ORIG_COMMIT_REV_ID: 5d5f6b7609fd7a1a995198f2cfc24cef78c864e7