[KGEN] Allow ParamOpInterface to specify how its parameter uses are scanned (#8720)
This PR adds a pair of methods to `ParamOpInterface` that allow
implementors to specify how the operation should be scanned for
parameters. This allows certain operations, particularly those that
exist at the boundary of two parameter scopes, to be explicit in which
parameter uses are in the current scope or in the defined scope.

MODULAR_ORIG_COMMIT_REV_ID: 559a7844c185fbeb113438bdd9fccbf2cd48358b