[KGEN] Fix a bunch of bugs in `always-inline-param` and enable it (#8781)
This PR fixes a ton of random bugs/holes in the `always-inline-param`
pass and enables the pass by default in the compilation pipeline.

It's definitely not the most performant pass in the world: parameter
mangling involves a lot of traversals and hash maps. In subsequent PRs,
I will dig into the performance of the pass.

MODULAR_ORIG_COMMIT_REV_ID: f5ca88c371489e99926c06cb54f07aae33604ebb