[lit-lang] Fix unqualified lookup of struct members to apply params. (#8862)
When performing unqualified lookup of a symbol and finding a method
in a struct, make sure to pick up the parameter bindings for the
enclosing struct.  For example, consider:

```
struct X[p1: Int]:
  @staticmethod
  fn someMethod[p2: Int](): pass

  fn thing():
    someMethod # <-- here
```

When looking up someMethod for unqualified lookup, include the
bound parameter value that is already in scope for 'p1', eliminating
the need (or ability) to redundantly specify it.

If you'd like to get access to 'someMethod' with a different binding
for the parameter, the correct way to do that is through qualified
lookup, e.g. `X[32].someMethod`.

This fixes #7576 and cleans up some code.

MODULAR_ORIG_COMMIT_REV_ID: c02c9ccbf04ea2aa4148dbe2f431cc5f7f6cdd4a