[lit-lang] Add type checking for a new `__clone__` member. (#8864)
This member corresponds to a "copy constructor" in C++, indicating
how to copy a value from one variable to another value.  For trivial
types, this is typically a bitwise copy, but for more complex types
this can call malloc or whatever is needed to create a logical copy
of the type.  The absence of this member will indicate non-copyable
types (like atomic variables).

This patch introduces type checking logic for this member, and adds
a couple implementations for existing types, but does not use it for
anything yet.  This is part of Issue #8599

MODULAR_ORIG_COMMIT_REV_ID: ecb15a36a61e639427cbf2e66b1aec61e2cead0a