[lit-lang] Implement a `Self` literal which returns the enclosing struct type. (#8865)
This is syntactic sugar for referring to the enclosing type with all
its parameter bound.  When we have classes, this will stop being
syntactic sugar and refer to the dynamic type of self instead, as
per PEP673.

We may eventually want to relax this to be a magic identifier some
day instead of a keyword, but using a keyword keeps things simple
and clean, which is the right place to start out from.

MODULAR_ORIG_COMMIT_REV_ID: a07d3aad4a4bc8d389c3461c9ac74b514b8e57dd