[lit-lang] Expand emitRValue to invoke `__clone__` during lv->rv promotion. (#8868)
Enhance the "loading" logic that  converts an lvalue to an rvalue:
when the underlying type is a raw MLIR type, continue to produce a
pop.load.  Otherwise, require that the user defined type implement
a `__clone__` method that implements the copy operation from the
lvalue to produce a fresh rvalue.

This is a big step towards implementing Issue #8599, but there is
one more piece needed as a follow-on.

MODULAR_ORIG_COMMIT_REV_ID: eb60886b675c8e6f04c5966b02d41bce282e89ec