[lit-lang] Introduce a new `@__memory_primary` struct decorator. (#8999)
This decorator currently just sets a bit on the litstructdecl, but
will be used to model whether a value is pervasively required to be
referenced in memory (e.g. on the stack) or whether it can be passed
around in SSA values.

I expect to flip the default for this bit when the model matures:
things should default to being memory primary and opt-in to being
passed by register.  This will take some time to build out though,
and this avoids breaking mainline development while this iterates.

MODULAR_ORIG_COMMIT_REV_ID: d22ca3036b0fabf102c9cc6d06251f1c803c2633