[lit-lang] Refactor an fix binop emission. (#9025)
It turns out that Python evaluates the RHS before an LHS in an
assignment statement, but the LHS before the RHS in an inplace
operator.  Adjust to match.  While here, do not pass in a contextual
type to inplace ops like `+=` because they are not allowed to
implicitly declare things in Python.

MODULAR_ORIG_COMMIT_REV_ID: dbeac2caa087e7c8e4662def87e202cbc0b86c47