[KGEN] Non parametric Recursion (#8989)
Recursion was broken for two reasons:
1. If function A called A, LowerLit erroneously considered A a nested
function
2. State assumptions: The result parameters of the ExpansionTreeNode are
expected to be set in the function that updates kgen callsites. The
order of concretization of a non leaf is (a) find bindings (b) replace
call ops with a binding's symbol and (c) set return type of node based
on result types of the return statement. In the case of recursion step (b) is dependent on step (c)
occurring first since the callee is not a child of the caller (children
get resolved first)

MODULAR_ORIG_COMMIT_REV_ID: 2197b51f1605bcc6c1c7d709bb52577f47c66bce