[LLCL] `constexpr` way for getting type names (#9108)
In our implementation of type id mappings, we rely on
`llvm::getTypeName()` which is a runtime implementation (poor QoI) since
it can be done at compile-time and we know the template type at
compile-time. Leverage this to do some gymnastics to build the demangled
string at compile-time. It's a bit annoying to do in C++17, but could be
simplified with C++20.

Apply the compile-time type name in `TypeNameResolver`. Don't change the
internal `StringMap`s to use `std::string_view` instead of `StringRef`
yet, but we can do in a follow-up (I think using `StringMap` out of the
box may not Just Work).

`typeNameFor` only works for types right now. The things permitted in a
non-type template parameter are not yet supported, but that's OK since
we don't have a use case for them right now (and upstream LLVM doesn't
support them either).

MODULAR_ORIG_COMMIT_REV_ID: 0798a0fa3bab9299dba38e1cfb796ab337067aaa