[LLCL] Finish refactor to TypeID (#9104)
- StateSlots use TypeID instead of own machinery.
- Since typeID.getValueDestructor() is now 'free' (as in 'already paid
for'...), use it in StateSlot and avoid the vtable for the virtual dtor.
- Don't support sub-typing in AsyncValue::get<T>. The previous
implementation would happily hand back the payload as a T if it was any
non-final class, which seemed too generous. At least for now lock it
down to exact type equality.
- Nuke some dead code I forgot to cleanup in #9056.
- Use TypeID::registerType<T>() throughout just to cut down on the
conceptual overhead.
- Give a more helpful error message if try to get an unregistered type.

With this LLCL is templated-statics free, which should unblock moving to
libmodular.so to share the same runtime between framework-specific
libraries.

MODULAR_ORIG_COMMIT_REV_ID: a5403152a9603ec58d9b69ebf8edb6d5081abc8f