[KGEN] Add JIT'ed code debugging. (#9087)
This patch re-adds the support for debugging code generated by the JIT (for example, in the `lit` tool).

There are some interesting caveats here:
(1) LLDB does not recognize breakpoints in lit code (at least not on macOS) - so you can't set a breakpoint and stop on it. You can step *into* lit code, though (as long as you don't need a source view, it will give you asm right now).
(2) We have to explicitly register the llvm_orc_registerJITLoaderGDB* functions because we build LLVM with hidden visibility. This will be harder once the DebugObjectManagerPlugin adds support for macOS, because it doesn't use a JITDylib to look up those functions, it uses a sys::DynamicLibrary.

These limitations appear to be an issue with LLDB and/or the ORC, not with our code. `lli` seems to have the same issue in my testing.

MODULAR_ORIG_COMMIT_REV_ID: 55b9345e5938a1efd1b1ea91b6613864a38280b9