[KGEN] Fix print method on ExpansionTreeNode. (#9112)
Currently, in a recursive call chain of A -> B -> A, we run into an infinite print loop where we're printing bindings forever because the binding is never `this`. This patch simply changes the debug logging to only print a single level of bindings - so we don't print bindings of bindings.

MODULAR_ORIG_COMMIT_REV_ID: a02c6b9b99ec8604077d5380cc83e832413d6725