[KGEN] Add error case for infinite parametric recursion. (#9113)
This patch makes handling more explicit for the case where we don't have result params on a node. If there are none required, then we're golden. If they are required, then we *must* have resolved them already - the only case when that's untrue that I can think of is if we're in an infinite recursive loop. This patch adds a test case for that and ensures we have a nice error message too.

MODULAR_ORIG_COMMIT_REV_ID: 0cee9ed091c56099c896c9453d3e7affc805fd78