[Cache] Ensure we don't re-store something we already have. (#9163)
The cache's policy on receiving something new with an existing hash is to simply overwrite. This is reasonable *for the CAS layer* but not really the behavior we want in general. Inside of CacheOps, we should check to see if we have the body/constant before we try to cache it.

MODULAR_ORIG_COMMIT_REV_ID: 0a1c0b9280f6fb47923ff5da6ee894dbd0f50865