[KGEN] Delete interfaces (#9274)
This PR deletes GeneratorInterfaceOp and everything related to it. The
only thing of note is that the interface constraint deduction in
LowerLIT has been gutted. It might be worthwhile to revive some of this
logic in the `@adaptive` decorator. Meanwhile, I will move the generator
constraint reduction to a separate pass that runs after
`always-inline-param`, since it's useless until `assert_param` is
inlined.

MODULAR_ORIG_COMMIT_REV_ID: 4f0a7751546b17378066961d1764c16d98aac4e3