[LLCL] Task queue overflow profiling annotations (#9202)
If lit's batch_matmul uses the batch size for its number of
parallelForEach tasks then the LLCL task-list can easily overflow,
forcing tasks to be run 'immediately'.

On dlrm-v2 (batch size 2048, though only 41x41 interior result matrices)
this appears to cause a 100us slowdown c.f. enlarging the task-list
queue capacity to avoid overflow. However that effect disappears with
larger inner matrix sizes.

I suspect the slowdown is because the tasks are executed in a more
erratic order when the tast list overflows. N tasks working their way
through the batch are placed in the queue in order, but when the queue
fills task N+1 is run immediately. Then some more tasks are enqueued.
Then another later task is run, and so on. That's more cache pressure I
guess.

I goofed around with using runItems to clear tasks from the head of the
queue, but nothing jumped out from the timing noise. However I will
include a unit 'test' to exercise this case, and I changed the profiling
entries to append ".now" when this happens to make it more obvious in
the traces.

Meanwhile, it takes 200ns or so for each worker to loop onto its next
task. So the overhead of adding many finely sliced tasks needs to be
balanced with its actual workload.

MODULAR_ORIG_COMMIT_REV_ID: 83d9e777d42986a88469fc43c556a83653faf1a3