[KGEN] Add `kgen.variadic.append` op. (#9500)
This op can be helpful in meta-programming when we need to work with
lists (represented as kgen.variadic values). E.g. it should allow
representing a cross-product of two lists, which is one of the things we
might want for the search use-case.

MODULAR_ORIG_COMMIT_REV_ID: 973bc96f72858b25283e3fbbf46052be1196b031