[lit-lang] Reimplement overloaded function emission. ~NFC (#9572)
This patch reimplements support for overloaded function emission
to eliminate CallableValue.  This replaces it with a new value
type (overloaded RValue, `ORValue`) that represents an emitted
value that has multiple candidates with different types.

This change is much more natural and is a simpler representation
for the parser, but introduces the complexity that ORValue doesn't
have a type.  To help address this, this adds a new value
classification (concrete RValue, `CRValue`) to manager RValues with
a known type.  Converting to a CRValue forces ambiguity to be
resolved, producing a single candidate.  The new value hierarchy
looks like this:

```
AnyValue              <- Expr emitted to MLIR:
  LValue (Value)        <- ... with a runtime address.
  RValue                <- ... as an independent value
    ORValue (OverloadSet) <- ..with an unresolved overload set
    CRValue               <- ..with a resolved type
      SRValue (Value)     <- ..with a dynamic value loaded into SSA register
      MRvalue (Value)     <- ..with a dynamic value emitted into memory
      PRValue (TypedAttr) <- ..with a parameter value (known at compile time)
```

This is conveniently a better representation for adaptive, and
enables ValueDest plumbing through call emission to work correctly.

MODULAR_ORIG_COMMIT_REV_ID: 0502a453176f8ee8bbcde4c2ce4317573cfdc9ed