[KGEN] Parallelize `lower-kgen-to-pop` (#9579)
The only reason we weren't running this in parallel is because we wanted
`applyPatternsAndFoldGreedily` to visit the function itself. Just handle
the function separately, which lets us run it as a nested pass.

The pass is still slower than it should be and that warrants some
investigation, but this is an easy win.

MODULAR_ORIG_COMMIT_REV_ID: 35473f8cdea506e4a97b106715ea76e49d783bec