[Support][ADT] Remove `isNull()` from `SmartVariant` (#9622)
`isNull()` and `operator bool()` on `SmartVariant` is currently a
pitfall waiting for users to step in. In the case of non-pointer types
being held in the `SmartVariant`, `isNull()` always returns false. This
is not ideal from a user perspective as they cannot provide a hint for
what nullability means in the case where the backing storage is a
`std::variant`.

Narrow the APIs a bit in `SmartVariant` by removing `operator bool` and
`isNull()`. Fix up the use in `IRValue` to not rely on `isNull()` from
`SmartVariant` anymore. The rationale for this is the same reason why
there lacks an `operator bool` for `std::variant` in principle. Given
that we want a consistent API surface for callers whether they are
dealing with pointer or non-pointer types, I don't see a way to easily
support this functionality at all - so just remove it. This pushes the
nullability check to users to solve in a different way, but that's OK,
since they will have more information and context to do it properly.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: affc5da858f1e247895aaf7ea7f162af20c4c2eb