[lit-lang] Implement var type inference with ValueDest. (#9699)
Now that call emission has been thoroughly refactored and is
cooperative with ValueDest, we can use it for var type inference.
This will allow us to emit memory primary values directly into the
var definition when inferring the type of the var.

This is a step forward, but we currently don't actually do that,
because calls cannot return things by-passed-in-reference. Now we
can build that logic.

MODULAR_ORIG_COMMIT_REV_ID: 67b2869e29fbfcad0a0df3dc827fc49557fa9772