[mt] Allow executing multiple inputs (#9638)
The `mt` driver only supports one input right now. This makes it hard to
test
multi-tenancy support in the `mt` tool which is desirable as a first
step for
multi-tenancy in the inference engine as a whole.

Add support for `mt` tool with multiple inputs for the `execute`
command. For
now, multiple inputs are specified via a comma-separated string that is
split on
just in the `mt` tool. This means the common CL option of
`inputFilename` can
stay as-is and we do not have to cause a breaking change to downstream
tools.
An alternative approach which may be done in the future is to introduce
a new CL
option just for the `mt` tool to allow for a list of strings as the
inputs to
avoid the need to split the string and such.

This change is mostly a refactoring to represent the state/datum for
each input
and looping over all of the input filenames specified. For most
commands, this
is just a refactoring. However, for the `execute` command, we don't want
to
stop after we run the binary file since we have multiple things to
execute
potentially. So, just keep executing them in the specified order. The
other
options to mt are assumed to be supported by all input types (such as
input
alignment size, shapes, etc.). This isn't an issue for running the same
model
on different frameworks, which is the current use case.

Add a multi-tenancy test that showcases the `mt` tool executing the same
model
for both TensorFlow and TFLite frameworks.

Note:
- In the future, we can add support for multiple output filenames so
each input
could have an associated output, and other fancy things. But for now,
just
  keep things very simple and no frills.

MODULAR_ORIG_COMMIT_REV_ID: 2c507eec63eac47e2080a0bff7063e77bdbcc337