[LLCL] Remove `TypeNameResolver` (#9691)
We don't have a need for this customization point of `TypeNameResolver`
yet, so remove it to simplify the code for getting a pretty compile-time
string representation of a type.

We can always reintroduce the customization point if needed later.

MODULAR_ORIG_COMMIT_REV_ID: cf12d6ffb2721d7d57bfce3111bdbb663b434dbf