[lit-lang] Teach function decls/calls/returns about memory-primary results. (#9735)
This allocates a return slot for types that must be returned in
registers,
and directly emplaces the returned value into that slot. This changes
the
call site to provide the value destination as the return slot, wiring up
the start of a big piece of support for memory primary types.

MODULAR_ORIG_COMMIT_REV_ID: 33015ed56acef5aa188e8c1cb804fb42cff154fe