[Kernels] Tame search space in `vectorize`. (#9654)
[Kernels] Tame search space in `vectorize`.

This is similar to #9626: in `vectorize[fn]` we have two calls:
`fn[simd_width]` and `fn[1]`. The former one is used in the main
vectorized loop and the latter one is used for the scalar remainder. We
do not need to search for the best possible option for `fn[1]` since its
effect on performance is small. Having search for it means that if `fn`
had `N` candidates, `vectorize[fn]` would have `N*N` candidates, which
doesn't scale well.

ALGORITHM_ORIG_REV_ID: 10e8a7826ad9415daf254b7eb265cbdca1115dc9