[lit-lang] Fix call result representation for memory-primary values. NFC. (#9752)
When calling a function that returns a memory-primary value, the
emitIR hook should return an MRValue for the result, not a PRValue
for the None that the function returns at the machine level.

This is NFC given the state of the stack.

MODULAR_ORIG_COMMIT_REV_ID: eafb6835985cc92c26b7d7614e8045a1e0039a0b