[lit-lang] Refactor the ValueDest passing-around model, NFC. (#9762)
It is imperative that expression emission emit into a ValueDest if
and when generating valid code (it is fine to ignore it when
generating an error).  This is easy enough to forget that we need
to have a way to check that it doesn't get forgotten on some path.
Right now the error manifests by silently failing to generate a
store, which is very difficult to debug.

This patch changes the convention for ValueDest getting passed
around: it now passes it down the stack by mutable reference, and
emitting to a ValueDest clears it out.  This will (in a future
patch) enable us to assert that the ValueDest has been correctly
emitted to and crash the compiler early on a bug.

MODULAR_ORIG_COMMIT_REV_ID: 990a7fbb1eddb25b037ebeae69c43efb15b2e413