[lit-lang] Implement support for mem-primary `let` decls. (#9779)
This required a fairly significant rework of how let processing
works: in the case of `let x = foo()` we won't know whether the
type of 'x' is register primary until after `foo()` is emitted.

We handle this by unifing the emission of let/var completely and
resolving the ambiguity late after the declaration is resolved,
promoting to a `lit.letreg.decl` after this is complete.  This
has the benefit of unifying a lot of code between var/let as well.

This patch also improves handling of rvalues, making 'x.y' produce
an RValue instead of an LValue when x is a memory private let,
teaches emitResult how to handle MRValues etc, all required to get
the new tests to work.

MODULAR_ORIG_COMMIT_REV_ID: 5021be0e87cdc647149bd7624f71860a1b2a3388