[KGEN] Refactor inlineGeneratorCall, NFC. (#9603)
This refactors inlineGeneratorCall to inline starting from a provided call op. The function is also exposed in a header file so we can use it from the elaborator in follow-up patches.

This exposed an interesting bug, and during debugging I discovered that we redeclare parameters in lit code sometimes! The symptom was that `struct<simd<size2, type2>>` resolved to `struct<scalar<type2>>` even though `size2` was 2 in the current scope. Because this patch is unruly enough, I've put in a simple fix, but we should investigate the cause of this more thoroughly: see [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 640c01f2983a1c09ceadfe3f82c045a1778cab61