[KGEN] Add KGEN_RegionAttr. (#9646)
This attribute is used in the elaborator to denote a bound region parameter. This simply refers directly back to the region operation itself, we can use the ParameterUseDefGraph to look up the region attached to this attribute (and this is what the next patch does).

MODULAR_ORIG_COMMIT_REV_ID: 3faf6c3c3275e5ec6ac6b2fd05a09a15eaff6329